+++ SeriesTracker/SeriesTracker/Dialogs/AddUserListDialog.xaml
@@ -2,7 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:wpf="clr-namespace:MaterialDesignThemes.Wpf;assembly=MaterialDesignThemes.Wpf"
-             xmlns:valid="clr-namespace:SeriesTracker.Controls"
+             xmlns:validators="clr-namespace:SeriesTracker.Utilities.ValidationRules"
+             xmlns:converters="clr-namespace:SeriesTracker.Utilities.Converters"
              xmlns:system="clr-namespace:System;assembly=mscorlib">
 	<Grid Margin="16">
 		<Grid.RowDefinitions>
@@ -14,8 +15,8 @@
 			<TextBox.Text>
 				<Binding Path="NewUserList" UpdateSourceTrigger="PropertyChanged">
 					<Binding.ValidationRules>
-						<valid:BadCharsValidationRule ValidatesOnTargetUpdated="True" />
-						<valid:NotEmptyValidationRule x:Name="ValidBadChars" ValidatesOnTargetUpdated="True" />
+						<validators:BadCharsValidationRule ValidatesOnTargetUpdated="True" />
+						<validators:NotEmptyValidationRule x:Name="ValidBadChars" ValidatesOnTargetUpdated="True" />
 					</Binding.ValidationRules>
 				</Binding>
 			</TextBox.Text>
@@ -28,7 +29,7 @@
 				<Button.IsEnabled>
 					<Binding ElementName="txt_Name" Path="(Validation.HasError)">
 						<Binding.Converter>
-							<valid:BoolToInverseBoolConverter />
+							<converters:BoolToInverseBoolConverter />
 						</Binding.Converter>
 					</Binding>
 				</Button.IsEnabled>
