+++ SeriesTracker/SeriesTracker/Dialogs/ProgressDialog/ProgressDialog.xaml.cs
