+++ SeriesTracker/SeriesTracker/Dialogs/ProgressDialog/ProgressDialogCancellationExcpetion.cs
