+++ SeriesTracker/SeriesTracker/Dialogs/ProgressDialog/ProgressDialogContext.cs
