+++ SeriesTracker/SeriesTracker/Dialogs/ProgressDialog/ProgressDialogResult.cs
