+++ SeriesTracker/SeriesTracker/Dialogs/ProgressDialog/ProgressDialogSettings.cs
