+++ SeriesTracker/SeriesTracker/Utilities/Commands/WPFCommandImplementation.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Input;
 
-namespace SeriesTracker.Controls
+namespace SeriesTracker.Utilities.Commands
 {
 	public class WPFCommandImplementation : ICommand
 	{
