+++ SeriesTracker/SeriesTracker/Windows/WindowMainNew.xaml
@@ -2,7 +2,7 @@
 	x:Class="SeriesTracker.Windows.WindowMainNew" 
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
-	xmlns:local="clr-namespace:SeriesTracker"
+	xmlns:converters="clr-namespace:SeriesTracker.Utilities.Converters"
 	xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
 	xmlns:ViewModels="clr-namespace:SeriesTracker.ViewModels"
 	Title="{Binding MyTitle}"
@@ -14,7 +14,7 @@
 		<ViewModels:MainNewViewModel/>
 	</Window.DataContext>
 	<Window.Resources>
-		<local:VisibilityToBoolConverter x:Key="VisibilityToBoolConv"/>
+		<converters:VisibilityToBoolConverter x:Key="VisibilityToBoolConv"/>
 	</Window.Resources>
 
 	<materialDesign:DialogHost Identifier="RootDialog" SnackbarMessageQueue="{Binding ElementName=MainSnackbar, Path=MessageQueue}">
