+++ SeriesTracker/SeriesTracker/Windows/WindowMainNew.xaml.cs
@@ -1,10 +1,10 @@
 ﻿using MaterialDesignThemes.Wpf;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using SeriesTracker.Comparers;
 using SeriesTracker.Core;
 using SeriesTracker.Dialogs;
 using SeriesTracker.Models;
+using SeriesTracker.Utilities.Comparers;
 using SeriesTracker.ViewModels;
 using System;
 using System.Collections.Generic;
